research_to_podcast/requirements.txt
anthropic>=0.49.0
gradio>=5.0.0
python-dotenv>=1.0.0
edge-tts>=6.1.9
pydub>=0.25.1
